# rpc, internal/guide: speed up tests a bit