This should remove about ~~two minutes~~ `75s` on CI. 

The `internal/guide` test was news to me, but apparently it executes the commands on the page here: https://geth.ethereum.org/docs/dapp/native-accounts . Well, doing full scrypt on CI is very slow, so maybe we don't have to do it all the time.  


On ` Image: Visual Studio 2019; Environment: GETH_ARCH=386, GETH_MINGW=C:\msys64\mingw32 `, without this PR 

```
ok  	github.com/ethereum/go-ethereum/internal/guide	18.755s	coverage: [no statements]
...
ok  	github.com/ethereum/go-ethereum/rpc	107.314s	coverage: 84.1% of statements
```
With this PR

```
ok  	github.com/ethereum/go-ethereum/internal/guide	1.751s	coverage: [no statements]
...
ok  	github.com/ethereum/go-ethereum/rpc	49.667s	coverage: 83.7% of statements
```