@holiman 2022-11-16
These are the biggest timesinks on the windows CI

<google-sheets-html-origin><style type="text/css"><!--td {border: 1px solid #cccccc;}br {mso-data-placement:same-cell;}--></style>

package | time
-- | --
github.com/ethereum/go-ethereum/tests | 320.869
github.com/ethereum/go-ethereum/eth/downloader | 150.764
github.com/ethereum/go-ethereum/p2p/discover | 99.935
github.com/ethereum/go-ethereum/trie | 82.508
github.com/ethereum/go-ethereum/consensus/ethash | 77.82
github.com/ethereum/go-ethereum/les | 69.552
github.com/ethereum/go-ethereum/eth/protocols/snap | 62.451
github.com/ethereum/go-ethereum/les/downloader | 58.636
github.com/ethereum/go-ethereum/core | 58.187
github.com/ethereum/go-ethereum/rpc | 49.667
github.com/ethereum/go-ethereum/eth/fetcher | 44.512
github.com/ethereum/go-ethereum/les/fetcher | 40.346
github.com/ethereum/go-ethereum/core/vm | 36.163
github.com/ethereum/go-ethereum/cmd/devp2p/internal/ethtest | 34.392
github.com/ethereum/go-ethereum/les/vflux/server | 34.363
github.com/ethereum/go-ethereum/cmd/geth | 23.341
github.com/ethereum/go-ethereum/core/state | 19.69
github.com/ethereum/go-ethereum/core/txpool | 17.696
github.com/ethereum/go-ethereum/les/flowcontrol | 17.407
github.com/ethereum/go-ethereum/metrics | 16.932
github.com/ethereum/go-ethereum/core/rawdb | 16.279
github.com/ethereum/go-ethereum/miner | 15.707
github.com/ethereum/go-ethereum/eth | 14.822
github.com/ethereum/go-ethereum/crypto/bls12381 | 13.906
github.com/ethereum/go-ethereum/accounts/keystore | 13.885